on avance

+++ graphik-manga/src/Entity/Products.php
@@ -20,7 +20,7 @@ class Products
     private ?string $taille = null;
 
     #[ORM\Column]
-    private ?int $prix = null;
+    private ?float $prix = null;
 
     #[ORM\Column(length: 255, nullable: true)]
     private ?string $public_cible = null;
@@ -60,12 +60,12 @@ class Products
         return $this;
     }
 
-    public function getPrix(): ?int
+    public function getPrix(): ?float
     {
         return $this->prix;
     }
 
-    public function setPrix(int $prix): static
+    public function setPrix(float $prix): static
     {
         $this->prix = $prix;
